construct constellation in universe from constellation data

+++ Assets/Scripts/Telescope/ConstellationGame/ConstellationStructure.cs
@@ -12,5 +12,5 @@ public class ConstellationStructure :ScriptableObject
     
     public Dictionary<int, List<int>> connections;
 
-    public int scale;
+    public float scale;
 }
